Start on maybe using more interfaces
I don't know if this is the right idea. It seems like a huge amount of
excessive complication for almost no gain, but who knows. Another thing
I'm wondering is if I should see the relationships between people,
wallets, and credit cards as hard or flexible, which would change how I
arrange things. I'm going to think on it for awhile and come back in a
bit.